requirements.txt
# API测试框架依赖
pytest==7.4.3
pytest-html==4.1.1
allure-pytest==2.13.2
requests==2.31.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pyyaml==6.0.1
python-dotenv==1.0.0
jsonschema==4.20.0
jsonpath-ng==1.6.0
jinja2==3.1.2
